change config's MaxConcurrent and janusItem's numSess type from int to uint64
because the config parser don't support int

+++ videoroom/janusHeap.go
@@ -11,7 +11,7 @@ import (
 
 type janusItem struct {
     janusConn   *janus.Janus
-    numSess      int
+    numSess      uint64
 }
 
 type janusHeap []*janusItem
